requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
idna==3.4
joblib==1.2.0
jwcrypto==1.4.2
numpy==1.24.1
oauthlib==3.2.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tzdata==2022.2
urllib3==1.26.13
wrapt==1.14.1
